requirements.txt
bcrypt==3.2.0
CacheControl==0.12.6
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
ecdsa==0.14.1
fastapi==0.63.0
firebase-admin==4.5.3
google-api-core==1.26.3
google-api-python-client==2.2.0
google-auth==1.28.1
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.1.0
google-cloud-storage==1.37.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.37.0
h11==0.12.0
httplib2==0.19.1
httptools==0.1.1
idna==2.10
msgpack==1.0.2
packaging==20.9
passlib==1.7.4
pprintjson==1.4.2
proto-plus==1.18.1
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.1
Pygments==2.8.1
pyparsing==2.4.7
python-dotenv==0.17.0
python-jose==3.2.0
python-multipart==0.0.5
python-slugify==4.0.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
starlette==0.13.6
text-unidecode==1.3
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
uvicorn==0.13.4
uvloop==0.15.2
watchgod==0.7
websockets==8.1
bcrypt==3.2.0
CacheControl==0.12.6
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
ecdsa==0.14.1
fastapi==0.63.0
firebase-admin==4.5.3
google-api-core==1.26.3
google-api-python-client==2.2.0
google-auth==1.28.1
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.1.0
google-cloud-storage==1.37.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.37.0
h11==0.12.0
httplib2==0.19.1
httptools==0.1.1
idna==2.10
msgpack==1.0.2
packaging==20.9
passlib==1.7.4
pprintjson==1.4.2
proto-plus==1.18.1
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.1
Pygments==2.8.1
pyparsing==2.4.7
python-dotenv==0.17.0
python-jose==3.2.0
python-multipart==0.0.5
python-slugify==4.0.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
starlette==0.13.6
text-unidecode==1.3
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
uvicorn==0.13.4
uvloop==0.15.2
watchgod==0.7
websockets==8.1
